requirements.txt
python-dotenv
werkzeug==2.1.2
Flask~=2.1.2
flask-restx==0.5.1
uWSGI
uwsgi-tools
click<8.1.0
Flask-Limiter
requests
qbittorrent-api==2023.6.50
Flask-APScheduler==1.12.4
git+https://github.com/Jorman/autoremove-torrents@patch-1
requests